x86_64: added a test for a function that uses _THIS_IP_
Commit 19b0aba67257 ("PPC64le - do not use the patched functions as callbacks directly")
in kpatch inadvertently fixed an issue on x86_64 as well.

If a non-static function was patched and that function used _THIS_IP_ somehow,
e.g. by calling some tracing functions, create-diff-object would fail
with an error like this:

  meminfo.o: changed function: meminfo_proc_show_ub
  create-diff-object: ERROR: meminfo.o:
  kpatch_create_intermediate_sections: 2485:
  lookup_local_symbol meminfo_proc_show_ub needed for .text.meminfo_proc_show_ub

Here is a patch to demonstrate this issue:
----------------
fs/proc/meminfo.c:
 @@ -95,6 +95,8 @@ int meminfo_proc_show_ub(struct seq_file *m, void *v,
 	unsigned long pages[NR_LRU_LISTS];
 	int lru;

 +	pr_info("[DBG] IP=%p\n", (void *)_THIS_IP_);
 +
 	si_meminfo(&i);
 	si_swapinfo(&i);

----------------

The problem was introduced by my commit
495e6197508d ("kpatch-build, x86: do not use the patched functions as callbacks directly").

Added a test with a function with _THIS_IP_ to avoid future regressions.

Signed-off-by: Evgenii Shatokhin <eshatokhin@virtuozzo.com>