Add unit test support for jump label support
For dynup/kpatch#937, the JUMP_STRUCT_SIZE environment variable is now
required, so add a hard-coded value to support it.

There's a potential future wrinkle here -- for recent upstream kernels,
starting with 50ff18ab497a ("jump_label: Implement generic support for
relative references"), the size of the jump table is no longer 24 bytes.
So for any new unit tests, they'll either have to be built against older
kernels, or we'll have to change the unit test makefiles to somehow
support different struct sizes.

But this should work for now.

Signed-off-by: Josh Poimboeuf <jpoimboe@redhat.com>